renamed Exception.hpp to ConfigException.hpp, modified adding error codes to existing targets and cleaned old code
