+++ config.conf
@@ -3,7 +3,7 @@ server {
     server_name 127.0.0.1;
 
     error_page 404 /custom_404.html;
-    error_page 502 503 503 /custom_50x.html;
+    error_page 502 503 504 /custom_50x.html;
     # error_page;
     # error_page 302;
