+++ inc/ConfigException.hpp
